+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/notification/domain/base/model/Notification.java
@@ -40,6 +40,7 @@ import java.util.Optional;
 
 import static java.util.stream.Collectors.groupingBy;
 import static org.apache.commons.collections4.ListUtils.emptyIfNull;
+import static org.eclipse.tractusx.traceability.common.date.DateUtil.convertInstantToString;
 
 @Data
 @Builder(toBuilder = true)
@@ -51,26 +52,32 @@ public class Notification {
     private NotificationStatus notificationStatus;
     private String description;
     private Instant createdAt;
+    private Instant updatedDate;
+
     private NotificationSide notificationSide;
     private NotificationType notificationType;
     @Builder.Default
     private List<String> affectedPartIds = new ArrayList<>();
-    private String closeReason;
-    private String acceptReason;
-    private String declineReason;
+    private NotificationSeverity severity;
+    private String targetDate;
+
     @Getter
     @Builder.Default
     private List<NotificationMessage> notifications = List.of();
 
 
-    public static Notification startNotification(String title, Instant createDate, BPN bpn, String description, NotificationType notificationType) {
+    public static Notification startNotification(String title, Instant createDate, BPN bpn, String description, NotificationType notificationType, NotificationSeverity severity, Instant targetDate) {
+
         return Notification.builder()
                 .title(title)
                 .bpn(bpn)
                 .notificationStatus(NotificationStatus.CREATED)
                 .notificationSide(NotificationSide.SENDER)
                 .notificationType(notificationType)
+                .targetDate(convertInstantToString(targetDate))
+                .severity(severity)
                 .description(description)
+                .updatedDate(Instant.now())
                 .createdAt(createDate)
                 .affectedPartIds(Collections.emptyList())
                 .build();
@@ -89,8 +96,6 @@ public class Notification {
                     applicationBPN,
                     editNotification.getReceiverBpn(),
                     editNotification.getDescription(),
-                    editNotification.getTargetDate(),
-                    editNotification.getSeverity(),
                     this.notificationType,
                     receiverAssetsMap,
                     applicationBPN.value(),
@@ -111,8 +116,6 @@ public class Notification {
                                 applicationBPN,
                                 sentToBPN,
                                 editNotification.getDescription(),
-                                editNotification.getTargetDate(),
-                                editNotification.getSeverity(),
                                 this.notificationType,
                                 receiverAssetsMapEntry,
                                 applicationBPN.value(),
@@ -134,38 +137,45 @@ public class Notification {
 
     public void cancel(BPN applicationBpn) {
         validateBPN(applicationBpn);
+        updatedDate = Instant.now();
         changeStatusTo(NotificationStatus.CANCELED);
-        this.closeReason = "canceled";
     }
 
-    public void close(BPN applicationBpn, String reason) {
+    public void close(BPN applicationBpn, String reason, NotificationMessage notificationMessage) {
         validateBPN(applicationBpn);
+        updatedDate = Instant.now();
         changeStatusTo(NotificationStatus.CLOSED);
-        this.closeReason = reason;
-        this.notifications.forEach(notification -> notification.setDescription(reason));
+        notificationMessage.setMessage(reason);
+        this.notifications.forEach(notification -> notification.setMessage(reason));
     }
 
     public void acknowledge() {
         changeStatusTo(NotificationStatus.ACKNOWLEDGED);
+        updatedDate = Instant.now();
+
     }
 
-    public void accept(String reason) {
+    public void accept(String reason, NotificationMessage message) {
         changeStatusTo(NotificationStatus.ACCEPTED);
-        this.acceptReason = reason;
+        updatedDate = Instant.now();
+        message.setMessage(reason);
     }
 
-    public void decline(String reason) {
+    public void decline(String reason, NotificationMessage message) {
         changeStatusTo(NotificationStatus.DECLINED);
-        this.declineReason = reason;
+        updatedDate = Instant.now();
+        message.setMessage(reason);
     }
 
-    public void close(String reason) {
+    public void close(String reason, NotificationMessage notificationMessage) {
         changeStatusTo(NotificationStatus.CLOSED);
-        this.closeReason = reason;
+        updatedDate = Instant.now();
+        notificationMessage.setMessage(reason);
     }
 
     public void send(BPN applicationBpn) {
         validateBPN(applicationBpn);
+        updatedDate = Instant.now();
         changeStatusTo(NotificationStatus.SENT);
     }