+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/common/support/IrsApiSupport.java
@@ -113,6 +113,16 @@ public class IrsApiSupport {
         );
     }
 
+    public void irsApiReturnsPoliciesNotFound(String id) {
+        whenHttp(restitoProvider.stubServer()).match(
+                Condition.get("/irs/policies/" +id)
+        ).then(
+                Action.status(HttpStatus.NOT_FOUND_404),
+                Action.header("Content-Type", "application/json"),
+                restitoProvider.jsonResponseFromFile("./stubs/irs/policies/response_404_get_policyById.json")
+        );
+    }
+
     public void irsApiReturnsPolicyById(String policyId) {
         whenHttp(restitoProvider.stubServer()).match(
                 Condition.get("/irs/policies/" + policyId)
@@ -130,6 +140,15 @@ public class IrsApiSupport {
                 Action.status(HttpStatus.OK_200),
                 Action.header("Content-Type", "application/json"),
                 restitoProvider.jsonResponseFromFile("./stubs/irs/policies/response_200_createPolicy.json")
+        );
+    }
+    public void irsApiCreatesPolicyBadRequest() {
+        whenHttp(restitoProvider.stubServer()).match(
+                Condition.post("/irs/policies")
+        ).then(
+                Action.status(HttpStatus.BAD_REQUEST_400),
+                Action.header("Content-Type", "application/json"),
+                restitoProvider.jsonResponseFromFile("./stubs/irs/policies/response_400_createPolicy.json")
         );
     }